[feat] #30 마이피드 무한스크롤

+++ app/src/main/java/com/android/go/sopt/winey/data/model/remote/response/ResponseGetWineyFeedListDto.kt
@@ -29,7 +29,7 @@ data class ResponseGetWineyFeedListDto(
             @SerialName("feedImage")
             val feedImage: String,
             @SerialName("feedMoney")
-            val feedMoney: Int,
+            val feedMoney: Long,
             @SerialName("feedTitle")
             val feedTitle: String,
             @SerialName("isLiked")
