+++ app/src/main/java/com/android/go/sopt/winey/domain/entity/WineyFeed.kt
@@ -4,7 +4,7 @@ package com.android.go.sopt.winey.domain.entity
 data class WineyFeed (
     val feedId: Int,
     val feedImage: String,
-    val feedMoney: Int,
+    val feedMoney: Long,
     val feedTitle: String,
     val isLiked: Boolean,
     val likes: Int,
